+++ src/main/java/dev/fxcte/creepyware/mixin/mixins/MixinGuiNewChat.java
@@ -1,7 +1,5 @@
 package dev.fxcte.creepyware.mixin.mixins;
 
-import java.awt.Color;
-import java.util.List;
 import dev.fxcte.creepyware.CreepyWare;
 import dev.fxcte.creepyware.features.modules.client.Colors;
 import dev.fxcte.creepyware.features.modules.misc.ChatModifier;
@@ -16,6 +14,9 @@ import org.spongepowered.asm.mixin.Shadow;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Redirect;
 
+import java.awt.*;
+import java.util.List;
+
 @Mixin(value={GuiNewChat.class})
 public class MixinGuiNewChat
 extends Gui {
@@ -26,7 +27,7 @@ extends Gui {
 
     @Redirect(method={"drawChat"}, at=@At(value="INVOKE", target="Lnet/minecraft/client/gui/GuiNewChat;drawRect(IIIII)V"))
     private void drawRectHook(int left, int top, int right, int bottom, int color) {
-        Gui.drawRect((int)left, (int)top, (int)right, (int)bottom, (int)(ChatModifier.getInstance().isOn() && ChatModifier.getInstance().clean.getValue() != false ? 0 : color));
+        Gui.drawRect(left , top , right , bottom , ChatModifier.getInstance().isOn() && ChatModifier.getInstance().clean.getValue() != false ? 0 : color);
     }
 
     @Redirect(method={"drawChat"}, at=@At(value="INVOKE", target="Lnet/minecraft/client/gui/FontRenderer;drawStringWithShadow(Ljava/lang/String;FFI)I"))
